fix(nextjs): Fix nextjs package to support new DSL
