+++ apps/api/syncs/entities/index.ts
@@ -1,3 +1,2 @@
-export * from './sync';
 export * from './sync_run';
 export * from './sync_run_mapper';
