+++ apps/api/temporal/sdk/customer_integrations/index.ts
@@ -1,5 +1,4 @@
-import { SyncConfig } from '@supaglue/types';
-import { Sync } from '../../../syncs/entities';
+import { Sync, SyncConfig } from '@supaglue/types';
 import {
   createDestinationSalesforce,
   createSourceSalesforce,
