+++ apps/api/temporal/sdk/customer_integrations/salesforce.ts
@@ -1,8 +1,7 @@
-import { SalesforceDestination, SalesforceSource, SyncConfig } from '@supaglue/types';
+import { SalesforceDestination, SalesforceSource, Sync, SyncConfig } from '@supaglue/types';
 import { ApplicationFailure } from '@temporalio/client';
 import retry from 'async-retry';
 import * as jsforce from 'jsforce';
-import { Sync } from '../../../syncs/entities';
 import { getMapping, getSalesforceObject, mapInternalToCustomerRecords } from '../../lib';
 import { BaseCustomerIntegration } from './base';
