+++ packages/nextjs/src/elements/FieldMapping/FieldMapping.tsx
@@ -1,24 +1,40 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
+import {
+  CustomerFieldMapping,
+  Field,
+  isOutboundSyncConfig,
+  isPostgresDestination,
+  PostgresDestination,
+  Schema,
+  SyncConfig,
+  SyncUpdateParams,
+} from '@supaglue/types';
 import classNames from 'classnames';
 import { ReactNode, useState } from 'react';
 import useSWR from 'swr';
 import useSWRMutation from 'swr/mutation';
 import { updateSync, useSalesforceIntegration } from '../../hooks/api';
-import {
-  CustomerFieldMapping,
-  DeveloperConfig,
-  Field,
-  PostgresDestination,
-  SyncConfig,
-  SyncUpdateParams,
-} from '../../lib/types';
 import { Select, SelectElements } from '../../primitives';
 import { XIcon } from '../../primitives/icons';
 import { SupaglueProviderInternal } from '../../providers';
 import { useSupaglueContext } from '../../providers/SupaglueProvider';
 import { SupaglueAppearance } from '../../types';
 import styles from './styles';
+
+const getSchema = (syncConfig: SyncConfig): Schema => {
+  if (isOutboundSyncConfig(syncConfig)) {
+    return syncConfig.source.schema;
+  }
+  return syncConfig.destination.schema;
+};
+
+const customPropertiesEnabled = (syncConfig: SyncConfig): boolean => {
+  if (isOutboundSyncConfig(syncConfig)) {
+    return !!syncConfig.source.config.customPropertiesColumn;
+  }
+  return isPostgresDestination(syncConfig.destination) && !!syncConfig.destination.config.customPropertiesColumn;
+};
 
 type MappedField = {
   name: string;
@@ -33,24 +49,20 @@ const EmptyContent = ({ className, children }: { className?: string; children: R
   );
 };
 
-const FieldMappingInternal = ({ appearance, syncConfigName }: FieldMappingProps) => {
+const FieldMappingInternal = ({ appearance, syncConfig }: FieldMappingProps) => {
   const { customerId } = useSupaglueContext();
   const { data: integration, error, isLoading: isLoadingIntegration } = useSalesforceIntegration(customerId);
 
-  const { data: developerConfig, isLoading: isLoadingDeveloperConfig } = useSWR<DeveloperConfig>({
-    path: '/developer_config',
-  });
-
   // TODO: Use conditional fetching syntax
   const { data: sync, isLoading: isLoadingSync } = useSWR({
-    path: `/syncs?customerId=${customerId}&syncConfigName=${syncConfigName}`,
+    path: `/syncs?customerId=${customerId}&syncConfigName=${syncConfig.name}`,
   });
 
   if (error) {
     return null;
   }
 
-  if (isLoadingDeveloperConfig || isLoadingSync || isLoadingIntegration) {
+  if (isLoadingSync || isLoadingIntegration) {
     return <EmptyContent>Loading...</EmptyContent>;
   }
 
@@ -62,17 +74,11 @@ const FieldMappingInternal = ({ appearance, syncConfigName }: FieldMappingProps)
     return <EmptyContent>No sync found.</EmptyContent>;
   }
 
-  if (!developerConfig?.syncConfigs?.length) {
-    return <EmptyContent>No developer config found.</EmptyContent>;
-  }
-
-  const syncConfig = developerConfig.syncConfigs.find(({ name }) => name === syncConfigName);
-
-  if (!syncConfig?.destination.schema.fields.length) {
+  if (!getSchema(syncConfig)) {
     return <EmptyContent>No fields to map.</EmptyContent>;
   }
 
-  return <FieldCollection appearance={appearance} key={syncConfigName} sync={sync} syncConfig={syncConfig} />;
+  return <FieldCollection appearance={appearance} key={syncConfig.name} sync={sync} syncConfig={syncConfig} />;
 };
 
 type FieldCollectionProps = {
@@ -90,10 +96,11 @@ type FieldCollectionProps = {
 };
 
 const FieldCollection = ({ appearance, syncConfig, sync }: FieldCollectionProps) => {
+  const schema = getSchema(syncConfig);
   // Use the customer-defined field mapping if it exists, or else the default one supplied by the developer
   const initialFieldMapping =
     sync.fieldMapping ??
-    syncConfig.destination.schema.fields.reduce<CustomerFieldMapping>((mapping, { name }) => {
+    schema.fields.reduce<CustomerFieldMapping>((mapping, { name }) => {
       mapping[name] = syncConfig.defaultFieldMapping?.find((entry) => entry.name === name)?.field || '';
       return mapping;
     }, {});
@@ -114,7 +121,7 @@ const FieldCollection = ({ appearance, syncConfig, sync }: FieldCollectionProps)
 
   const { upsertKey } = (syncConfig.destination as PostgresDestination).config;
 
-  const onUpdateSync = async (params: SyncUpdateParams, onSuccess?: () => void) => {
+  const onUpdateSync = async (id: string, params: SyncUpdateParams, onSuccess?: () => void) => {
     const result = await callUpdateSync(params);
 
     if (result?.data) {
@@ -133,21 +140,16 @@ const FieldCollection = ({ appearance, syncConfig, sync }: FieldCollectionProps)
     };
 
     await onUpdateSync(
+      sync.id,
       {
-        id: sync.id,
+        type: syncConfig.type,
         fieldMapping: updatedFieldMapping,
       },
       () => setFieldMapping(updatedFieldMapping)
     );
   };
 
-  // TODO: Add flag to enable/disable custom properties
-  // For now, just check to see if there's a column configured for custom properties
-  const destinationConfig = syncConfig.destination.config;
-  const customPropertiesEnabled =
-    'customPropertiesColumn' in destinationConfig && !!destinationConfig.customPropertiesColumn;
-
-  const applicationFields = [...syncConfig.destination.schema.fields, ...(sync.customProperties || [])];
+  const applicationFields = [...schema.fields, ...(sync.customProperties || [])];
   const customPropertyNames = new Set((sync.customProperties || []).map((field) => field.name));
 
   const onCreateCustomProperty = async (name: string) => {
@@ -160,8 +162,9 @@ const FieldCollection = ({ appearance, syncConfig, sync }: FieldCollectionProps)
     }
 
     await onUpdateSync(
+      sync.id,
       {
-        id: sync.id,
+        type: syncConfig.type,
         customProperties: [...(sync.customProperties || []), { name, label: name }],
       },
       () => setIsCreatingCustomProperty(false)
@@ -173,8 +176,9 @@ const FieldCollection = ({ appearance, syncConfig, sync }: FieldCollectionProps)
     delete fieldMapping[name];
 
     await onUpdateSync(
+      sync.id,
       {
-        id: sync.id,
+        type: syncConfig.type,
         customProperties: (sync.customProperties || []).filter((field) => field.name !== name),
         fieldMapping: updatedFieldMapping,
       },
@@ -239,7 +243,9 @@ const FieldCollection = ({ appearance, syncConfig, sync }: FieldCollectionProps)
         />
       )}
 
-      {customPropertiesEnabled && (
+      {/* TODO: Add flag to enable/disable custom properties
+      For now, just check to see if there's a column configured for custom properties */}
+      {customPropertiesEnabled(syncConfig) && (
         <AddCustomPropertyButton
           appearance={appearance}
           disabled={isCreatingCustomProperty}
@@ -323,11 +329,11 @@ type FieldMappingAppearance = SupaglueAppearance & {
 
 export type FieldMappingProps = {
   appearance?: FieldMappingAppearance;
-  syncConfigName: string;
+  syncConfig: SyncConfig;
 };
 
-export const FieldMapping = ({ appearance, syncConfigName }: FieldMappingProps) => (
+export const FieldMapping = ({ appearance, syncConfig }: FieldMappingProps) => (
   <SupaglueProviderInternal>
-    <FieldMappingInternal appearance={appearance} syncConfigName={syncConfigName} />
+    <FieldMappingInternal appearance={appearance} syncConfig={syncConfig} />
   </SupaglueProviderInternal>
 );
